Add dry-run before Question for apply/destroy/run
